+++ resources/aws-apigateway-deployment_stagedescription.go
@@ -22,7 +22,7 @@ type AWSApiGatewayDeployment_StageDescription struct {
 	// CacheTtlInSeconds AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apitgateway-deployment-stagedescription.html#cfn-apigateway-deployment-stagedescription-cachettlinseconds
-	CacheTtlInSeconds int64 `json:"CacheTtlInSeconds"`
+	CacheTtlInSeconds int `json:"CacheTtlInSeconds"`
 
 	// CachingEnabled AWS CloudFormation Property
 	// Required: false
@@ -67,7 +67,7 @@ type AWSApiGatewayDeployment_StageDescription struct {
 	// ThrottlingBurstLimit AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apitgateway-deployment-stagedescription.html#cfn-apigateway-deployment-stagedescription-throttlingburstlimit
-	ThrottlingBurstLimit int64 `json:"ThrottlingBurstLimit"`
+	ThrottlingBurstLimit int `json:"ThrottlingBurstLimit"`
 
 	// ThrottlingRateLimit AWS CloudFormation Property
 	// Required: false
