+++ resources/aws-apigateway-usageplan_throttlesettings.go
@@ -7,7 +7,7 @@ type AWSApiGatewayUsagePlan_ThrottleSettings struct {
 	// BurstLimit AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apigateway-usageplan-throttlesettings.html#cfn-apigateway-usageplan-throttlesettings-burstlimit
-	BurstLimit int64 `json:"BurstLimit"`
+	BurstLimit int `json:"BurstLimit"`
 
 	// RateLimit AWS CloudFormation Property
 	// Required: false
