+++ resources/aws-autoscaling-launchconfiguration_blockdevice.go
@@ -17,7 +17,7 @@ type AWSAutoScalingLaunchConfiguration_BlockDevice struct {
 	// Iops AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-as-launchconfig-blockdev-template.html#cfn-as-launchconfig-blockdev-template-iops
-	Iops int64 `json:"Iops"`
+	Iops int `json:"Iops"`
 
 	// SnapshotId AWS CloudFormation Property
 	// Required: false
@@ -27,7 +27,7 @@ type AWSAutoScalingLaunchConfiguration_BlockDevice struct {
 	// VolumeSize AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-as-launchconfig-blockdev-template.html#cfn-as-launchconfig-blockdev-template-volumesize
-	VolumeSize int64 `json:"VolumeSize"`
+	VolumeSize int `json:"VolumeSize"`
 
 	// VolumeType AWS CloudFormation Property
 	// Required: false
