+++ resources/aws-batch-jobdefinition_containerproperties.go
@@ -27,7 +27,7 @@ type AWSBatchJobDefinition_ContainerProperties struct {
 	// Memory AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-batch-jobdefinition-containerproperties.html#cfn-batch-jobdefinition-containerproperties-memory
-	Memory int64 `json:"Memory"`
+	Memory int `json:"Memory"`
 
 	// MountPoints AWS CloudFormation Property
 	// Required: false
@@ -57,7 +57,7 @@ type AWSBatchJobDefinition_ContainerProperties struct {
 	// Vcpus AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-batch-jobdefinition-containerproperties.html#cfn-batch-jobdefinition-containerproperties-vcpus
-	Vcpus int64 `json:"Vcpus"`
+	Vcpus int `json:"Vcpus"`
 
 	// Volumes AWS CloudFormation Property
 	// Required: false
