+++ resources/aws-batch-jobdefinition_retrystrategy.go
@@ -7,7 +7,7 @@ type AWSBatchJobDefinition_RetryStrategy struct {
 	// Attempts AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-batch-jobdefinition-retrystrategy.html#cfn-batch-jobdefinition-retrystrategy-attempts
-	Attempts int64 `json:"Attempts"`
+	Attempts int `json:"Attempts"`
 }
 
 // AWSCloudFormationType returns the AWS CloudFormation resource type
