+++ resources/aws-batch-jobdefinition_ulimit.go
@@ -7,7 +7,7 @@ type AWSBatchJobDefinition_Ulimit struct {
 	// HardLimit AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-batch-jobdefinition-ulimit.html#cfn-batch-jobdefinition-ulimit-hardlimit
-	HardLimit int64 `json:"HardLimit"`
+	HardLimit int `json:"HardLimit"`
 
 	// Name AWS CloudFormation Property
 	// Required: true
@@ -17,7 +17,7 @@ type AWSBatchJobDefinition_Ulimit struct {
 	// SoftLimit AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-batch-jobdefinition-ulimit.html#cfn-batch-jobdefinition-ulimit-softlimit
-	SoftLimit int64 `json:"SoftLimit"`
+	SoftLimit int `json:"SoftLimit"`
 }
 
 // AWSCloudFormationType returns the AWS CloudFormation resource type
