+++ resources/aws-cloudfront-distribution_customoriginconfig.go
@@ -7,12 +7,12 @@ type AWSCloudFrontDistribution_CustomOriginConfig struct {
 	// HTTPPort AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-cloudfront-customorigin.html#cfn-cloudfront-customorigin-httpport
-	HTTPPort int64 `json:"HTTPPort"`
+	HTTPPort int `json:"HTTPPort"`
 
 	// HTTPSPort AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-cloudfront-customorigin.html#cfn-cloudfront-customorigin-httpsport
-	HTTPSPort int64 `json:"HTTPSPort"`
+	HTTPSPort int `json:"HTTPSPort"`
 
 	// OriginProtocolPolicy AWS CloudFormation Property
 	// Required: true
