+++ resources/aws-codepipeline-pipeline_actiondeclaration.go
@@ -37,7 +37,7 @@ type AWSCodePipelinePipeline_ActionDeclaration struct {
 	// RunOrder AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-codepipeline-pipeline-stages-actions.html#cfn-codepipeline-pipeline-stages-actions-runorder
-	RunOrder int64 `json:"RunOrder"`
+	RunOrder int `json:"RunOrder"`
 }
 
 // AWSCloudFormationType returns the AWS CloudFormation resource type
