+++ resources/aws-ec2-instance_ebs.go
@@ -17,7 +17,7 @@ type AWSEC2Instance_Ebs struct {
 	// Iops AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-blockdev-template.html#cfn-ec2-blockdev-template-iops
-	Iops int64 `json:"Iops"`
+	Iops int `json:"Iops"`
 
 	// SnapshotId AWS CloudFormation Property
 	// Required: false
@@ -27,7 +27,7 @@ type AWSEC2Instance_Ebs struct {
 	// VolumeSize AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-blockdev-template.html#cfn-ec2-blockdev-template-volumesize
-	VolumeSize int64 `json:"VolumeSize"`
+	VolumeSize int `json:"VolumeSize"`
 
 	// VolumeType AWS CloudFormation Property
 	// Required: false
