+++ resources/aws-ec2-networkaclentry_icmp.go
@@ -7,12 +7,12 @@ type AWSEC2NetworkAclEntry_Icmp struct {
 	// Code AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-networkaclentry-icmp.html#cfn-ec2-networkaclentry-icmp-code
-	Code int64 `json:"Code"`
+	Code int `json:"Code"`
 
 	// Type AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-networkaclentry-icmp.html#cfn-ec2-networkaclentry-icmp-type
-	Type int64 `json:"Type"`
+	Type int `json:"Type"`
 }
 
 // AWSCloudFormationType returns the AWS CloudFormation resource type
