+++ resources/aws-ec2-securitygroup_ingress.go
@@ -17,7 +17,7 @@ type AWSEC2SecurityGroup_Ingress struct {
 	// FromPort AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-security-group-rule.html#cfn-ec2-security-group-rule-fromport
-	FromPort int64 `json:"FromPort"`
+	FromPort int `json:"FromPort"`
 
 	// IpProtocol AWS CloudFormation Property
 	// Required: true
@@ -42,7 +42,7 @@ type AWSEC2SecurityGroup_Ingress struct {
 	// ToPort AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-security-group-rule.html#cfn-ec2-security-group-rule-toport
-	ToPort int64 `json:"ToPort"`
+	ToPort int `json:"ToPort"`
 }
 
 // AWSCloudFormationType returns the AWS CloudFormation resource type
